Added requirements

+++ api/models.py
@@ -19,7 +19,7 @@ class Wallet(models.Model):
 class MpesaTransaction(models.Model):
     txn_id = models.CharField(max_length=128, )
     mpesa_receipt = models.CharField(max_length=32, blank=True, null=True,default="")
-    amount = models.PositiveIntegerField()
+    amount = models.DecimalField(decimal_places=2, max_digits=9)
     reason = models.CharField(max_length=128, )
     account = models.CharField(max_length=128, default="")
     TXN_TYPES = (
